Remove [id] from path for schedule-visit
And redirect to it.
